+++ examples/forms-bench/CHANGELOG.md
@@ -1,5 +1,12 @@
 # glove-forms-bench
 
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-memory@1.2.1
+
 ## 1.0.4
 
 ### Patch Changes