+++ examples/foundry-agent/CHANGELOG.md
@@ -1,5 +1,16 @@
 # glove-foundry-example
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-memory@1.2.1
+  - glove-foundry@0.3.3
+  - glove-js@0.4.3
+  - glove-working-environment@0.6.1
+
 ## 1.0.3
 
 ### Patch Changes