+++ examples/foundry-braind-storm/CHANGELOG.md
@@ -1,5 +1,17 @@
 # glove-foundry-braind-storm
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-voice-s2s@0.2.3
+  - glove-foundry@0.3.3
+  - glove-image@0.2.2
+  - glove-mesh@0.1.4
+  - glove-working-environment@0.6.1
+
 ## 1.0.3
 
 ### Patch Changes