+++ examples/glovebox-smoke/CHANGELOG.md
@@ -1,5 +1,12 @@
 # glovebox-smoke
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glovebox-core@0.6.3
+
 ## 0.0.3
 
 ### Patch Changes