+++ examples/layered-voice/CHANGELOG.md
@@ -1,5 +1,16 @@
 # glove-layered-voice
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-voice-s2s@0.2.3
+  - glove-mesh@0.1.4
+  - glove-voice@3.6.3
+  - glove-next@3.1.4
+
 ## 1.0.3
 
 ### Patch Changes