+++ examples/s2s-rooms/CHANGELOG.md
@@ -1,5 +1,14 @@
 # glove-s2s-rooms
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-voice-s2s@0.2.3
+  - glove-mesh@0.1.4
+
 ## 1.0.3
 
 ### Patch Changes