+++ examples/scratchpad-agent/CHANGELOG.md
@@ -1,5 +1,13 @@
 # glove-scratchpad-example
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-scratchpad@2.0.3
+
 ## 1.0.3
 
 ### Patch Changes