+++ packages/glove-egress/CHANGELOG.md
@@ -1,5 +1,12 @@
 # glove-egress
 
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-scratchpad@2.0.3
+
 ## 0.2.2
 
 ### Patch Changes