+++ packages/glove-lisp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # glove-lisp
 
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - glove-core@3.8.0
+  - glove-scratchpad@2.0.3
+
 ## 0.4.2
 
 ### Patch Changes